# Conflicts:
#	balancebot/common/exchanges/binance/binance.py

+++ balancebot/common/models/ticker.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import NamedTuple
 
 
@@ -5,5 +6,5 @@ class Ticker(NamedTuple):
 
     symbol: str
     exchange: str
-    price: float
+    price: Decimal
     ts: int
