+++ balancebot/common/models/trade.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import NamedTuple
 
 from balancebot.common.enums import Side
@@ -8,8 +9,8 @@ class Trade(NamedTuple):
 
     symbol: str
     side: str
-    size: float
-    price: float
+    size: Decimal
+    price: Decimal
     exchange: str
     time: datetime
     perp: bool
